refactor : converter 패키지 삭제 -> service/utils 로 이전
