+++ src/main/java/prgrms/marco/be02marbox/domain/theater/service/TheaterService.java
@@ -3,10 +3,10 @@ package prgrms.marco.be02marbox.domain.theater.service;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import prgrms.marco.be02marbox.domain.theater.service.utils.TheaterConverter;
 import prgrms.marco.be02marbox.domain.theater.Theater;
 import prgrms.marco.be02marbox.domain.theater.dto.RequestCreateTheater;
 import prgrms.marco.be02marbox.domain.theater.repository.TheaterRepository;
+import prgrms.marco.be02marbox.domain.theater.service.utils.TheaterConverter;
 
 @Service
 public class TheaterService {
