+++ src/test/java/prgrms/marco/be02marbox/domain/theater/service/TheaterServiceTest.java
@@ -11,11 +11,11 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.context.annotation.Import;
 
-import prgrms.marco.be02marbox.domain.theater.service.utils.TheaterConverter;
 import prgrms.marco.be02marbox.domain.theater.Region;
 import prgrms.marco.be02marbox.domain.theater.Theater;
 import prgrms.marco.be02marbox.domain.theater.dto.RequestCreateTheater;
 import prgrms.marco.be02marbox.domain.theater.repository.TheaterRepository;
+import prgrms.marco.be02marbox.domain.theater.service.utils.TheaterConverter;
 
 @DataJpaTest
 @Import({TheaterService.class, TheaterConverter.class})
